fix(ui): split pure neuron formatters out of neuron-table.tsx to fix bundle budget
movers-band.tsx (rendered on the homepage) imports taoCompact from
neuron-table.tsx, which now also imports StakeUnstakeModal's wallet-connect
chain (#6219). That pulled the whole staking flow into the homepage's
initial bundle, pushing it to 303 KB against the 300 KB gzip budget and
breaking main's Validate/ui check on every commit since.

Moves taoCompact, scoreStr, and SponsoredBadge into a new neuron-format.tsx
with no heavy imports, and repoints every consumer at it. NeuronTable itself
is unchanged. No visual/behavioral change — verified locally: bundle back
to 293 KB, lint/format/typecheck/test (1023/1023) all green.
